role check done

+++ app/Http/Controllers/Admin/AboutUsController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\RoleEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\AboutUs\StoreAboutUsRequest;
 use App\Http\Requests\Admin\AboutUs\UpdateAboutUsRequest;
+use App\Repositories\Product\ProductRepository;
 use App\Services\Admin\AboutUsService;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
@@ -20,6 +22,12 @@ use Illuminate\Support\Facades\Session;
 
 class AboutUsController extends Controller
 {
+
+    public function __construct(ProductRepository $productRepository)
+    {
+        parent::__construct($productRepository);
+    }
+
     /**
      * @param AboutUsService $service
      * @param Request $request
@@ -27,6 +35,9 @@ class AboutUsController extends Controller
      */
     public function index(AboutUsService $service): Response
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getAllData();
         viewShare($response);
         return response()->view('admin.about-us.index');
@@ -39,6 +50,9 @@ class AboutUsController extends Controller
      */
     public function create(AboutUsService $service): Response
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getCreateData();
 
         viewShare($response);
@@ -52,6 +66,9 @@ class AboutUsController extends Controller
      */
     public function store(AboutUsService $service, StoreAboutUsRequest $request): RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->addNewData($request->validated());
 
         if ($this->isError($response)) return $this->getErrorResponse();
@@ -67,6 +84,9 @@ class AboutUsController extends Controller
      */
     public function edit(AboutUsService $service, int $id): Response|RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getEditData($id);
         if ($this->isError($response)) return $this->getErrorResponse();
 
@@ -82,6 +102,9 @@ class AboutUsController extends Controller
      */
     public function update(AboutUsService $service, UpdateAboutUsRequest $request, int $id): RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->updateDataById($id, $request->validated());
         if ($this->isError($response)) return $this->getErrorResponse();
 
@@ -96,6 +119,9 @@ class AboutUsController extends Controller
      */
     public function destroy(AboutUsService $service, int $id):RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->deleteDataById($id);
         if ($this->isError($response)) return $this->getErrorResponse();
