+++ app/Http/Controllers/Admin/UsersController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Enums\RoleEnum;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Users\StoreUserRequest;
 use App\Http\Requests\Admin\Users\UpdateUserRequest;
@@ -26,6 +27,9 @@ class UsersController extends Controller
      */
     public function index(UserService $service):Response
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getAllDataPaginated();
         viewShare($response);
         return response()->view("admin.users.index");
@@ -37,6 +41,9 @@ class UsersController extends Controller
      */
     public function create(UserService $service):Response
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getCreateData();
 
         viewShare($response);
@@ -51,6 +58,9 @@ class UsersController extends Controller
      */
     public function store(UserService $service, StoreUserRequest $request):RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->addNewData($request->validated());
 
         if ($this->isError($response)) return $this->getErrorResponse();
@@ -67,6 +77,9 @@ class UsersController extends Controller
      */
     public function edit(UserService $service, int $id):Response|RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         $response = $service->getEditData($id);
 
         if ($this->isError($response)) return $this->getErrorResponse();
@@ -84,6 +97,9 @@ class UsersController extends Controller
      */
     public function update(UserService $service, UpdateUserRequest $request,int $id):RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         (array) $response = $service->updateDataById($id, $request->validated());
 
         if ($this->isError($response)) return $this->getErrorResponse();
@@ -99,6 +115,9 @@ class UsersController extends Controller
      */
     public function destroy(UserService $service, int $id):RedirectResponse
     {
+        if(!auth()->user()->hasRole(RoleEnum::ADMINISTRATOR())){
+            abort(403);
+        }
         (array) $response = $service->deleteDataById($id);
 
         if ($this->isError($response)) return $this->getErrorResponse();
